Renamed mock_factory module to mock_factories
