+++ halotools/empirical_models/mock_factories.py
