+++ halotools/empirical_models/test_empirical_models/__init__.py
@@ -8,5 +8,5 @@ from .test_halo_prof_param_components import *
 from .test_halo_prof_components import *
 from .test_halo_prof_components import *
 from .test_gal_prof_components import *
-from .test_mock_factory import *
+from .test_mock_factories import *
 from .test_hod_factory import *
